\docType{data}
\name{annotCarbayo}
\alias{annotCarbayo}
\title{Carbayo annotation  data}
\description{
  Contains annotations for 93 samples of Carbayo data.
\format{
  A data frame with sample annotations.
}
\author{
  Anne Biton
}
\references{
  \url{http://jco.ascopubs.org/content/24/5/778/suppl/DC1}
}
\keyword{datasets}

