\docType{data}
\name{dataCarbayo}
\alias{dataCarbayo}
\title{Carbayo expression data}
\description{
  Contains bladder cancer expression data based on on
  HG-U133A Affymetrix microarrays. The data include 93
  samples, were normalized with MAS5 by the authors of the
  paper using Quantile normalization and
  log2-transformation. They are restricted to the 10000
  most variable probe sets.
}
\author{
  Anne Biton
}
\references{
  \url{http://jco.ascopubs.org/content/24/5/778/suppl/DC1}
}
\keyword{datasets}

