\docType{data}
\name{icaSetStransky}
\alias{icaSetStransky}
\title{IcaSet-object containing a FastICA decomposition of gene expression microarrray-based data of bladder cancer samples.}
\description{
  Object of class \code{\link{IcaSet}} containing an ICA
  decomposition calculated by the FastICA algorithm
  (through matlab function "icasso") on bladder cancer
  expression data measured on HG-U133-95a and HG-U133-95av2
  Affymetrix microarrays. It contains 20 independent
  components. The original expression data contain 63 tumor
  samples and were normalized by RMA with
  log2-transformation.
}
\author{
  Anne Biton
}
\references{
  \url{http://microarrays.curie.fr/publications/oncologie_moleculaire/bladder_TCM/}
}
\keyword{datasets}

